Error and promise rejection events must be fired on the global object (#82)
As decided in the March 6th meeting, this PR specifies that the
`error`, `unhandledrejection` and `rejectionhandled` events are fired
on the global object. It also adds the `ErrorEvent` and
`PromiseRejectionEvent` interfaces, and the `onerror`,
`onunhandledrejection` and `onrejectionhandled` event handler
properties, to the minimum common API list.

This PR also has a note covering the likely case that Node.js will
not support this due to legacy.